adk10/pyproject.toml
[project]
name = "adk10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.24.1",
]
